fix bug where attempted_approval is sometimes true when it shouldn't be
If a match is queued to be posted, removed from the queue, then rejected, it would be marked as attempted_approval = true from being queued and then always stay true when explicitly rejected. It should not be the case for any match that rejected is true, attempted_approval is true, and auto_rejected is false. The only rejected attempted approvals are auto-rejections from one of the tweets in the match being inaccessible.